Assign maximum height as a float in the scans
The various arrays are all floats rather than doubles, so z should be
a float, not a double.

+++ src/morphology/reconstruct.c
@@ -56,7 +56,7 @@ ptrdiff_t forward_scan(float *marker, float *mask, ptrdiff_t nrows,
     // maximum height of the neighborhood and the mask at the current
     // pixel.
 
-    double z = max_height < mask[p] ? max_height : mask[p];
+    float z = max_height < mask[p] ? max_height : mask[p];
 
     if (z != marker[p]) {
       // Increment count only if we change the current pixel
@@ -121,7 +121,7 @@ ptrdiff_t backward_scan(float *marker, float *mask, ptrdiff_t nrows,
     // maximum height of the neighborhood and the mask at the current
     // pixel.
 
-    double z = max_height < mask[p] ? max_height : mask[p];
+    float z = max_height < mask[p] ? max_height : mask[p];
     if (z != marker[p]) {
       // Increment count only if we change the current pixel
       count++;
